refactor(cmd): move the CSV cap and ranking logic into its own file
cmd/multi_service.go carried the whole --input-csv cap path alongside the
purchase pipeline. Extract the functions this PR added for that path into
cmd/multi_service_csv_cap.go so the file stops growing against the 500-line
limit in CLAUDE.md on account of this change.

Moved verbatim: scoreAndLimitCSVRecs, applyMinCountFloor, savingsPerInstance,
sortBySavingsPerInstance, maxNamedUnrankableRows and requireRankingSignal.
Each is reached only from the CSV path: applyMinCountFloor from runToolFromCSV
and scoreAndLimitCSVRecs, the rest from scoreAndLimitCSVRecs alone or from one
another.

capBinds and the rankingRule constants stay behind despite being new here,
because applyGlobalInstanceLimit, scoreLimitAndDisplay and reportInstanceLimit
consume them on the recommendation-driven path too. The four pre-existing cap
helpers are untouched.

This is a pure move. The 148 moved lines are byte-identical, the diff on
cmd/multi_service.go is 0 additions and 151 deletions (the block, one blank
separator, and the sort/strings imports it took with it), and no signature,
name or behaviour changed. cmd/multi_service.go goes from 881 to 730 lines.
The remainder above its 698-line pre-PR size, and the wider file-size breach
across cmd/, is tracked in #1834.
